+++ R/utils.R
@@ -915,10 +915,13 @@ tweak_hugo_env = function(baseURL = NULL, relativeURLs = NULL, server = FALSE) {
   c2 = if (is.null(relativeURLs)) get_config('relativeurls', FALSE, config) else relativeURLs
   if (server && c1) b = paste0(if (grepl('^//', b)) 'http:' else 'http://example.org/', b)
 
-  v = set_envvar(c(
-    HUGO_BASEURL = if (c2) '/' else b, HUGO_RELATIVEURLS = tolower(c2),
-    BLOGDOWN_POST_RELREF = if (server) 'true' else NA
-  ))
+  vars = c(HUGO_BASEURL = if (c2) '/' else b, HUGO_RELATIVEURLS = tolower(c2))
+  if (server) {
+    vars = c(vars, BLOGDOWN_POST_RELREF = 'true')
+    c3 = get_config('ignoreErrors', NA, config)
+    if (is.na(c3)) vars = c(vars, HUGO_IGNOREERRORS = 'error-remote-getjson')
+  }
+  v = set_envvar(vars)
   exit_call(function() set_envvar(v))
 }